requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
psycopg2-binary>=2.9
openai>=1.0
python-dotenv>=1.0
Pillow>=10.0
